requirements.txt
fastapi
python-docx
uvicorn
python-multipart